requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
supabase>=2.10.0
openai>=1.54.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.27.0
python-multipart>=0.0.12

